Move GetFilesForDeployment to files module
