+++ cmd/sad/main_test.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"os"
-	"path/filepath"
 	"testing"
 
 	"github.com/jswny/sad"
@@ -127,57 +126,6 @@ func TestParseFlagsEmptyValues(t *testing.T) {
 	}
 
 	testutils.CompareOpts(testOpts, *opts, t)
-}
-
-func TestGetFilesForDeployment(t *testing.T) {
-	dirName := "dir.test"
-
-	tempDirPath, err := ioutil.TempDir("", dirName)
-
-	if err != nil {
-		t.Fatalf("Error creating temp dir: %s", err)
-	}
-
-	defer os.RemoveAll(tempDirPath)
-
-	fileNames := []string{
-		main.DockerComposeFileName,
-	}
-
-	content := []byte("test")
-
-	for _, fileName := range fileNames {
-		filePath := filepath.Join(tempDirPath, fileName)
-
-		if err := ioutil.WriteFile(filePath, content, 0755); err != nil {
-			t.Fatalf("Error writing to temp file \"%s\", %s", filePath, err)
-		}
-	}
-
-	files, err := main.GetFilesForDeployment(tempDirPath)
-
-	if err != nil {
-		t.Fatalf("Error getting files for deployment: %s", err)
-	}
-
-	expected := len(fileNames)
-	actual := len(files)
-
-	if actual != expected {
-		t.Errorf("Getting files for deployment returned %d files, expected %d", actual, expected)
-	}
-
-	for _, file := range files {
-		data, err := ioutil.ReadFile(file.Name())
-
-		if err != nil {
-			t.Fatalf("Error reading from deployment file: %s", err)
-		}
-
-		if string(content) != string(data) {
-			t.Errorf("Expected file content %s but got %s", content, data)
-		}
-	}
 }
 
 func buildArgs(stringOpts *testutils.StringOptions) []string {
