+++ files.go
@@ -2,9 +2,16 @@ package sad
 
 import (
 	"errors"
+	"fmt"
 	"os"
 	"path/filepath"
 )
+
+// DockerComposeFileName is the name of the Docker Compose file to be loaded for deployment.
+var DockerComposeFileName string = ".sad.docker-compose.yml"
+
+// DotEnvFileName is the name of the .env file to be loaded for deployment.
+var DotEnvFileName string = ".sad.env"
 
 // FindFilePathRecursiveFileNotFoundErrorMessage is the string error message returned when FindFilePathRecursive cannot find the specified file.
 var FindFilePathRecursiveFileNotFoundErrorMessage = "file not found"
@@ -34,4 +41,40 @@ func FindFilePathRecursive(fromPath string, fileName string) (string, error) {
 	}
 
 	return foundPath, nil
+}
+
+// GetFilesForDeployment gets and opens the files needed for deployment by finding them recursively under the provided fromPath.
+// Files: Docker Compose file (see DockerComposeFileName).
+// Opens files, remember to close.
+func GetFilesForDeployment(fromPath string) ([]*os.File, error) {
+	var filePaths []string
+	var files []*os.File
+
+	fileNames := []string{
+		DockerComposeFileName,
+	}
+
+	for _, fileName := range fileNames {
+		filePath, err := FindFilePathRecursive(fromPath, fileName)
+
+		if err != nil {
+			err := fmt.Errorf("error finding file \"%s\" under path \"%s\": %s", fileName, fromPath, err)
+			return nil, err
+		}
+
+		filePaths = append(filePaths, filePath)
+	}
+
+	for _, filePath := range filePaths {
+		file, err := os.Open(filePath)
+
+		if err != nil {
+			err := fmt.Errorf("error opening file for deployment from path \"%s\"", filePath)
+			return nil, err
+		}
+
+		files = append(files, file)
+	}
+
+	return files, nil
 }
